Replace arbitrary int with named constant
In definitions of custom tag fieldinfo'es, the arbitrary `-1` for `readCount` and `writeCount` has been replaced with the named constant `TiffFieldInfo.Variable`.  Note that I don't speak vb, so if the constant needs to be accessed by some other name (such as 'TiffFieldInfo.[Variable]'), you'll need to correct that separately or let me know in this push's comments (where I'll be happy to do so).