Added overloads for use cases that already cache directory info
For high performance applications which cache and already know the directory offsets, I added overloads to set a directory by setting the offset.

Also added method to get the count of directories in the file, this enables someone to ready number of pages in a tiff file by recursing the directories and optionally caching the info for fast lookup later.